+++ pumpkin/src/net/java/play.rs
@@ -21,7 +21,7 @@ use crate::log_at_level;
 use crate::net::PlayerConfig;
 use crate::net::java::JavaClient;
 use crate::plugin::block::block_place::BlockPlaceEvent;
-use crate::plugin::player::player_bucket_entity::PlayerBucketEntityEvent;
+use crate::plugin::player::bucket_entity::PlayerBucketEntityEvent;
 use crate::plugin::player::player_chat::PlayerChatEvent;
 use crate::plugin::player::player_command_send::PlayerCommandSendEvent;
 use crate::plugin::player::player_interact_entity_event::PlayerInteractEntityEvent;
