+++ pumpkin/src/plugin/api/events/player/bucket_empty.rs
